pyproject.toml
[tool.poetry]
name = "pythonproject2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"
pandas-stubs = "~=2.3.3"
streamlit = "^1.54.0"
plotly = "^6.5.2"
yfinance = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
